backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
torch>=2.0.0
transformers>=4.38.2
Pillow>=10.2.0
kaggle>=1.6.6
